third_party/blink/web_tests/platform/generic/svg/dynamic-updates/SVGFESpotLightElement-dom-pointsAtZ-attr-expected.txt

SVG 1.1 dynamic update tests

Tests dynamic updates of the 'pointsAtZ' attribute of the SVGFESpotLightElement object

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS spotLightElement.getAttribute('pointsAtZ') is "100"
PASS spotLightElement.getAttribute('pointsAtZ') is "0"
PASS successfullyParsed is true

TEST COMPLETE
